REFACTOR: UpdateAPI refactor
- Update API Refactored
